# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/security": {
      "get": {
        "operationId": "ApiController_security[0]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "resources"
        ]
      },
      "patch": {
        "operationId": "ApiController_updateSecurity[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecurityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "resources"
        ]
      }
    },
    "/security": {
      "get": {
        "operationId": "ApiController_security[1]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "resources"
        ]
      },
      "patch": {
        "operationId": "ApiController_updateSecurity[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecurityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "resources"
        ]
      }
    },
...
    "/api/v1/auth/email-verification/request": {
      "post": {
        "operationId": "SecurityFlowsController_requestEmail[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/email-verification/request": {
      "post": {
        "operationId": "SecurityFlowsController_requestEmail[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/email-verification/confirm": {
      "post": {
        "operationId": "SecurityFlowsController_confirmEmail[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/email-verification/confirm": {
      "post": {
        "operationId": "SecurityFlowsController_confirmEmail[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/pin/setup/verify": {
      "post": {
        "operationId": "SecurityFlowsController_pinSetup[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/pin/setup/verify": {
      "post": {
        "operationId": "SecurityFlowsController_pinSetup[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/2fa/enroll": {
      "post": {
        "operationId": "SecurityFlowsController_enroll[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/2fa/enroll": {
      "post": {
        "operationId": "SecurityFlowsController_enroll[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/2fa/confirm": {
      "post": {
        "operationId": "SecurityFlowsController_confirm[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/2fa/confirm": {
      "post": {
        "operationId": "SecurityFlowsController_confirm[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/2fa/login/verify": {
      "post": {
        "operationId": "SecurityFlowsController_login[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/2fa/login/verify": {
      "post": {
        "operationId": "SecurityFlowsController_login[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/2fa/disable": {
      "post": {
        "operationId": "SecurityFlowsController_disable[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/2fa/disable": {
      "post": {
        "operationId": "SecurityFlowsController_disable[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/biometric/registration/options": {
      "post": {
        "operationId": "SecurityFlowsController_registrationOptions[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/biometric/registration/options": {
      "post": {
        "operationId": "SecurityFlowsController_registrationOptions[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/biometric/registration/verify": {
      "post": {
        "operationId": "SecurityFlowsController_register[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/biometric/registration/verify": {
      "post": {
        "operationId": "SecurityFlowsController_register[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/step-up/biometric/challenge": {
      "post": {
        "operationId": "SecurityFlowsController_challenge[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BiometricChallengeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/step-up/biometric/challenge": {
      "post": {
        "operationId": "SecurityFlowsController_challenge[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BiometricChallengeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/auth/step-up/biometric": {
      "post": {
        "operationId": "SecurityFlowsController_verify[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/auth/step-up/biometric": {
      "post": {
        "operationId": "SecurityFlowsController_verify[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationProofDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/security/biometric-credentials": {
      "get": {
        "operationId": "SecurityFlowsController_credentials[0]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/security/biometric-credentials": {
      "get": {
        "operationId": "SecurityFlowsController_credentials[1]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/security/biometric-credentials/{id}": {
      "delete": {
        "operationId": "SecurityFlowsController_revokeCredential[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/security/biometric-credentials/{id}": {
      "delete": {
        "operationId": "SecurityFlowsController_revokeCredential[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/security/push-tokens": {
      "post": {
        "operationId": "SecurityFlowsController_pushToken[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/security/push-tokens": {
      "post": {
        "operationId": "SecurityFlowsController_pushToken[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/api/v1/security/push-tokens/{id}": {
      "delete": {
        "operationId": "SecurityFlowsController_removePush[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    },
    "/security/push-tokens/{id}": {
      "delete": {
        "operationId": "SecurityFlowsController_removePush[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "security flows"
        ]
      }
    }
  },
  "info": {
    "title": "Invento prototype API",
    "description": "Simulated finance. No real financial rails.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SecurityDto": {
        "type": "object",
        "properties": {
          "stepUpToken": {
            "type": "string"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "biometricsEnabled": {
            "type": "boolean"
          },
          "autoLock": {
            "type": "number"
          },
          "transactionAlerts": {
            "type": "boolean"
          },
          "loginAlerts": {
            "type": "boolean"
          }
        }
      },
...
      "ResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "TokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "PasswordProofDto": {
        "type": "object",
        "properties": {
          "stepUpToken": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
      "OtpDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "MfaLoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "challengeToken": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "challengeToken"
        ]
      },
      "DisableMfaDto": {
        "type": "object",
        "properties": {
          "stepUpToken": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "code"
        ]
      },
      "StepDto": {
        "type": "object",
        "properties": {
          "stepUpToken": {
            "type": "string"
          }
        }
      },
      "RegistrationProofDto": {
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string"
          },
          "response": {
            "type": "object"
          }
        },
        "required": [
          "challengeToken",
          "response"
        ]
      },
      "BiometricChallengeDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "transfer",
              "investment_order",
              "beneficiary_add",
              "pin_change",
              "password_change",
              "security_downgrade",
              "biometric_enroll",
              "two_factor_setup"
            ]
          }
        },
        "required": [
          "action"
        ]
      },
      "AuthenticationProofDto": {
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string"
          },
          "response": {
            "type": "object"
          }
        },
        "required": [
          "challengeToken",
          "response"
        ]
      },
      "PushTokenDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "expo",
              "fcm",
              "apns"
            ]
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "token"
        ]
      }
    }
  }
}